Updated Paths
Updated all the paths when working in the base library to not reference
folders. This will always force smaller bundles if a folder gains new
exports later down the road. It also is a bit safer when trying to
bundle things together.. Knowning exactly what is being brought it.

This also fixes it so all the tests don't failt atm either.