Fixed SelectFieldColumn and EditDialogColumn #170
The SelectFieldColumn was rewritten so that is positions itself
absolutely a bit more accurately so that it can correctly overflow in
data tables. Also updated the SelectFieldColumn and EditDialogColumn so
that when they are open, they will reposition themselves if the user
scrolls the table left or right.

To get a better rendering experience, I had to remove the `position:
relative` from the `.md-table-column`, so tooltips no longer work. I
will need to fix that before release.