Fixed double onChange firing for Switch #182
The switch's thumb and the switch's track need the onClick handler so
that either the track can be clicked or the thumb can be _clicked_ by
pressing the enter key. However, if a user actually touches or clicks
the thumb itself, the click event propigates from the thumb to the track
as well, so two onChange events are emitted.

This commit prevents that click propigation from the switch's thumb.