Updated Slider Precision with Floating Values
There was a problem when the slider had a step between 0 and 1 that the
discrete label would not display correctly when dragging - only when
incrementing with the text field or keyboard.

Extracted out the floating point vulnerablity for the #184 bugfix so it
could be tested separately since I suck at maths.

Added some simple tests for the calculateValueDistance function as well.