Added a better nested dialog fix
The initial "fix" for the nested dialog solution was really a cop-out of
laziness since I didn't feel like _really_ fixing it before 1.1.0's
Portal updates.

This fix will now automatically place components that use the Portal
component from within a Dialog to be rendered in the dialog itself
instead of the last or first child of the body. So now dialogs can be
nested within each other and now have to worry about z-indexes and all
that jazz.

This should _really_ close #229 now.