Updated Drawer/NavigationDrawer type resolving
Upated the logic for resolving the current drawer's media type to allow
the props.type to always be used instead of attempting to find the
current media type match.

This is really only useful when the preferred usage of the drawer is
to use one of the temporary types across on media devices.

Closes #291