Updated Accessibility for Form Controls
It loosk like I haven't played around with native form controls enough..
The AccessibleFakeButton will now correctly attempt to submit a form
when the enter key is pressed instead of triggering the click event.
This means that Checkbox, Radio, and Switch will now *not* be
toggle-able with the enter key. It will now require the space key.