SSR Fixes and Separate Stores
Ended up splitting the store into a seprate one for client and server.
The server will be very simple and not have sagas while the client will.

Added a way to fetch all the data before rendering the page as well by
creating a getInitialState function for the server.

Added the docgens state and aded fetches for it.

Updated how the redux dev tools work for the new version.