Fixed and simplified TextField Blur Logic
The TextField will no logner manually call the blur function when an
area outside of the text field is touched/clicked. It will fall back to
the browser/os defaults on how to handle blur events.

The active state for the text field is now triggered on the blur event
instead of the manual blur call. I can't remember why I implemented it
this way to begine with.

Closes #434