Updated Multiline TextFields for Height Calcs
The Mutliline text field will now correctly calculate its height if the
styles have loaded after initial render as well as if the page text
field is resized manually.

In addition, this fixes the bug that the multiline text field does not
render correctly when not in a flex container.

Closes #365