Fixed the Layover Positioning
When the Layover was toggled quickly between visible and invisible, it
sometimes incorrectly positioned the elements. This mostly happened
because the init function was called to set some assumed styles before
the child fully rendered. Once the child mounts, the ref callback was
called to finish the positioning.

This logic was flawed if the child never fully unmounted inbetween
because the final positioning was never called again. This fix stops the
init styling to be caclualted/applied if the Layover's visibility is
toggled in quick succession.

Closes #345
Closes #449 (I think?)