Context Menu Layover Fixes
Updated the logic for context menus so that it is a lot more accurate
at positioning. When there is no selected text, it will use the mouse
location. If there is selected text, it will use the center of the
selected text as the basepoint.

This also fixes the weird bug in Firefox that automatically closes
context menus. The height is still a bit off in FF, but it is reasonable
to me.

Closes #430