Fixed the SelectField Focus Logic
The SelectField will now correctly open the menu with a down key press
and then require one more down keypress to focus the first item in the
list.

When the activeItem is stripped, the logic for focusing has also been
fixed. It used to take 1 additional down keypress to focus the first
item.